Use int32 instead of long and uint32 instead of ulong, so less chances for int size mismatches on 64 bit

+++ ArpHeader/ArpCollections/ArpIndexedList.h
@@ -133,7 +133,7 @@ class ArpIndexedList {
 									// nn, the next element will have events
 									// from nn+1 to mm, etc.  When we need to
 									// add higher events, we grow the chain.
-		long		count;			// the size of the chain
+		int32		count;			// the size of the chain
 
 	private:
 //		ReadWriteLock	chainLock;
