+++ ArpSrc/AmKernel/AmTrackInfoView.cpp
@@ -208,7 +208,7 @@ void AmTrackInfoView::PostDrawSliceOn(BView* view, BRect clip)
 void AmTrackInfoView::HandleDraggedTrackData(BMessage *msg) {
 #if 0
 	BWindow		*win;
-	long		dragIndex /*, index = Index()*/;
+	int32		dragIndex /*, index = Index()*/;
 	if ( (msg->FindPointer(STR_WINDOW, (void**)&win) != B_OK)
 			|| (msg->FindInt32(STR_POSITION, &dragIndex) != B_OK) )
 		return;
